Data on decisions

This dataset include data on infringement decisions. There is one observation per decision per case (2002-2020). The dataset includes information on the department responsible for the decision and the member state that the decision is against.

decisions

Format

A data frame with 33 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

decision_id

Numeric. An ID number that uniquely identifies each decision in an infringement case.

case_number

Numeric. The number of the case.

case_year

Numeric. The year of the case.

decision_date

Date. The date that the Commission issued the decision in the format YYYY-MM-DD.

decision_year

Numeric. The year that the Commission issued the decision.

decision_month

Numeric. The month that the Commission issued the decision.

decision_day

Numeric. The day that the Commission issued the decision.

member_state_id

Numeric. An ID number that uniquely identifies each member state. This ID number is assigned when member states are sorted by accession date and then alphabetically.

member_state

String. The name of the member state that the Commission opened the case against.

member_state_code

String. A two letter code assigned by the Commission that uniquely identifies each member state.

department_id

Numeric. An ID number that uniquely identifies each Commission department.

department

String. The name of the Commission department that opened the infringement case.

department_code

String. A multi-letter code assigned by the Commission that uniquely identifies each department.

case_type_id

Numeric. An ID number that uniquely identifies each type of state aid cases. Coded 1 for noncommunication cases, which are cases that relate to a member state failing to notify the Commission that it has transposed a directive by the stated deadline. Coded 2 for nonconformity cases, which are cases that relate to a member state incorrectly transposing a directive.

case_type

String. The type of the infringement case. There are two types of cases. Coded noncommunication for cases that relate to a member state failing to notify the Commission that it has transposed a directive by the stated deadline. Coded nonconformity for cases that relate to a member state incorrectly transposing a directive.

noncommunication

Dummy. A dummy variable indicating whether the infringement case is a noncommunication case, which is a case that relates to a member state failing to notify that Commission that it has transposed a directive by the stated deadline.

nonconformity

Dummy. A dummy variable indicating whether the infringement case is a nonconformity case, which is a case that relates to a member state incorrectly transposing a directive.

directive

Dummy. A dummy variable indicating whether the description of the infringement case provided by the Commission references a directive that the case relates to. Directive numbers have the format ####/####, where the first set of numbers is the year the directive was enacted and the second set is the number of the directive.

directive_number

String. The number of the directive that the infringement case relates to, if applicable. Coded NA if the description of the case provided by the Commission does not reference a specific directive number.

celex

String. The CELEX number of the directive that the infringement case relates to, if applicable. Coded NA if the description of the case provided by the Commission does not reference a specific directive number. CELEX numbers for directives have the format 3####L####, where the first set of numbers is the year the directive was enacted and the second set is the number of the directive.

decision_stage_id

Numeric. An ID number that uniquely identifies each decision stage in the infringement procedure. Coded 1 for letters of formal notice under Article 258 of the Treaty on the Functioning of the European Union (TFEU), coded 2 for reasoned opinions under Article 258, coded 3 for referrals to the Court of Justice of the European Union (CJEU) under Article 258, coded 4 for letters of formal notice under Article 260, coded 5 for reasoned opinions under Article 260, and coded 6 for referrals to the Court under Article 260.

decision_stage

String. The decision stage of the infringement procedure. Possible values include: Letter of formal notice (Article 258), Reasoned opinion (Article 259), Referral to the Court (Article 258), Letter of formal notice (Article 260), Reasoned opinion (Article 260), and Referral to the Court (Article 260).

stage_lfn_258

Dummy. A dummy variable indicating whether the decision is a letter of formal notice sent by the Commission under Article 258 of the Treaty on the Functioning of the European Union (TFEU).

stage_ro_258

Dummy. A dummy variable indicating whether the decision is a reasoned opinion sent by the Commission under Article 258 of the Treaty on the Functioning of the European Union (TFEU).

stage_rf_258

Dummy. A dummy variable indicating whether the decision is a referral to the Court of Justice of the European Union (CJEU) made by the Commission under Article 258 of the Treaty on the Functioning of the European Union (TFEU).

stage_lfn_260

Dummy. A dummy variable indicating whether the decision is a letter of formal notice sent by the Commission under Article 260 of the Treaty on the Functioning of the European Union (TFEU).

stage_ro_260

Dummy. A dummy variable indicating whether the decision is a reasoned opinion sent by the Commission under Article 260 of the Treaty on the Functioning of the European Union (TFEU).

stage_rf_260

Dummy. A dummy variable indicating whether the decision is a referral to the Court of Justice of the European Union (CJEU) made by the Commission under Article 260 of the Treaty on the Functioning of the European Union (TFEU).

stage_closing

Dummy. A dummy variable indicating whether the decision is the closing of the case by the Commission.

stage_withdrawal

Dummy. A dummy variable indicating whether the decision is the withdrawal of the case by the Commission.

stage_additional

Dummy. A dummy variable indicating whether the decision is an additional decision (a decision of the same type as a previous decision).

press_release

Dummy. A dummy variable indicating whether the Commission published a press release to publicize the decision.