CSTS data on decisions (by department)

This dataset includes aggregated data on the number of decisions per stage of the infringement procedure per department per year (cross-sectional time-series data). There is one observation per department per year per decision stage (2002-2020). The dataset uses current department names.

decisions_csts_dp

Format

A data frame with 8 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

year

Numeric. The year the decision was issued by the Commission.

department_id

Numeric. An ID number that uniquely identifies each Commission department.

department

String. The name of the Commission department that opened the infringement case.

department_code

String. A multi-letter code assigned by the Commission that uniquely identifies each department.

decision_stage_id

Numeric. An ID number that uniquely identifies each decision stage in the infringement procedure. Coded 1 for letters of formal notice under Article 258 of the Treaty on the Functioning of the European Union (TFEU), coded 2 for reasoned opinions under Article 258, coded 3 for referrals to the Court under Article 258, coded 4 for letters of formal notice under Article 260, coded 5 for reasoned opinions under Article 260, and coded 6 for referrals to the Court under Article 265

decision_stage

String. The decision stage of the infringement procedure. Possible values include: Letter of formal notice (Article 258), Reasoned opinion (Article 259), Referral to the Court (Article 258), Letter of formal notice (Article 260), Reasoned opinion (Article 260), and Referral to the Court (Article 260).

count_decisions

Numeric. A count of the number of decisions made by the Commission in infringement cases at this level of aggregation.