Network data on decisions (by case type)

This dataset includes multi-dimensional network data for infringement decisions. There is one dimension per case type (noncommunication vs nonconformity) per decision stage. Network data is similar to directed dyad-year data except that it only includes directed dyad-years with at least one decision with respect to each case type and stage. For every year, there is one node per department and one node per member state. Edges can only exist between a department and a member state. There is an edge between a department and a member state with respect to each case type and stage if and only if the department issued at least one decision against the member state during that year. The weight of the edge is the number of decisions that the department opened against the member state. There is one observation per department per member state per year per case type per decision stage (2002-2020), excluding directed dyad-years where the state was not a member of the EU, but only if the count of decisions is positive.

decisions_net_d

Format

A data frame with 11 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

time

Numeric. The time period, where 2002 is coded 1.

d1_layer_id

Numeric. An ID number that uniquely identifies each layer of the network in the first dimension.

d1_layer

String. The layer of the network in the first dimension, which is type of case.

d2_layer_id

Numeric. An ID number that uniquely identifies each layer of the network in the second dimension.

d2_layer

String. The layer of the network in the second dimension, which is the decision stage.

from_node_id

Numeric. An ID number that uniquely identifies each node in the network that creates a link, which is always a Commission department.

from_node

String. The name of the Commission department that opened the infringement case.

to_node_id

Numeric. An ID number that uniquely identifies each node in the network that receives a link, which is always a member state.

to_node

String. The name of the member state that the Commission opened the case against.

edge_weight

Numeric. The weight of the edge, which is the number of decisions opened by the Commission.